[FE/style] MODIFY create_challenge_screen.dart (widget -> stateful class 로 변경)
