+++ frontend/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:frontend/challenge/create/create_challenge_screen.dart';
+import 'package:frontend/challenge/create/create_challenge_screen_fir.dart';
 import 'package:frontend/challenge/detail/detail_challenge_screen.dart';
 import 'package:frontend/challenge/state/state_challenge_screen.dart';
 import 'package:frontend/community/tab_community_screen.dart';
@@ -23,7 +24,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   @override
-  // This widget is the root of your application.
+  // This widgets is the root of your application.
   Widget build(BuildContext context) {
     return ScreenUtilInit(
         designSize: const Size(375, 844),
@@ -37,7 +38,7 @@ class _MyAppState extends State<MyApp> {
                 // SplashScreen.routeName: (context) => SplashScreen(),
                 'login': (context) => const LoginScreen(),
                 'main' : (context) => const MainScreen(),
-                'create_challenge' :  (context) => CreateChallenge(),
+                'create_challenge' :  (context) => CreateChallenge_fir(),
                 'detail_challenge' : (context) => ChallengeDetailScreen(),
                 'state_challenge' : (context) => ChallengeStateScreen(),
                 'community' : (context) => TabCommunityScreen(),
