+++ frontend/test/widget_test.dart
@@ -1,9 +1,9 @@
-// This is a basic Flutter widget test.
+// This is a basic Flutter widgets test.
 //
-// To perform an interaction with a widget in your test, use the WidgetTester
+// To perform an interaction with a widgets in your test, use the WidgetTester
 // utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
+// gestures. You can also use WidgetTester to find child widgets in the widgets
+// tree, read text, and verify that the values of widgets properties are correct.
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
